py_requirements.txt
beautifulsoup4==4.11.1
dnspython==2.2.1
pymongo==4.1.1
python-dotenv==0.20.0
requests==2.27.1
certifi==2021.10.8
